refactor: move domain mixins into superme_sdk/services/
_http.py stays at root (transport infra, not a service).
